+++ tests/test_optim_apsimx.R
@@ -92,7 +92,12 @@ if(FALSE){
   ## Run the model with incorrect parameters
   ## RUE = 1.2
   ## Phyllochron = 120
-  system.time(sim.b4 <- apsimx("Wheat-opt-ex.apsimx", src.dir = extd.dir))
+  tmp <- tempdir()
+  setwd(tmp)
+  file.copy(file.path(extd.dir, "Ames.met"), ".")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  
+  system.time(sim.b4 <- apsimx("Wheat-opt-ex.apsimx", src.dir = tmp)) ## This takes 4-5 seconds
   
   ## write.csv(sim.b4, "wheat-sim-b4-opt.csv", row.names = FALSE)
 
@@ -127,7 +132,6 @@ if(FALSE){
   ## This takes about ~7.5 minutes
   start <- Sys.time()
   wop <- optim_apsimx("Wheat-opt-ex.apsimx", 
-                       src.dir = extd.dir, 
                        parm.paths = c(pp1, pp2),
                        data = obsWheat, 
                        weights = "mean",
@@ -202,12 +206,16 @@ if(FALSE){
   
 }
 
+#### Classic ####
 if(FALSE){
 
+  tmp <- tempdir()
+  setwd(tmp)
+  
   apsim_options(warn.versions = FALSE)
   
   ## Testing optimization for an APSIM Classic example
-  ## Millet
+  ## Maize
   if(file.exists("Maize.apsim")) file.remove("Maize.apsim")
   if(file.exists("Maize.xml")) file.remove("Maize.xml")
   if(file.exists("Maize.out")) file.remove("Maize.out")
@@ -260,6 +268,14 @@ if(FALSE){
                      hessian = TRUE,
                      weights = "mean")
   
+  # op2.u <- optim_apsim("Maize.apsim", 
+  #                      crop.file = "Maize.xml",
+  #                      parm.paths = c(pp1, pp2),
+  #                      type = "ucminf",
+  #                      data = obsMaize,
+  #                      hessian = TRUE,
+  #                      weights = "mean")
+  
   op2nl <- optim_apsim("Maize.apsim", 
                      crop.file = "Maize.xml",
                      parm.paths = c(pp1, pp2),
@@ -303,4 +319,135 @@ if(FALSE){
   if(file.exists("Maize.xml")) file.remove("Maize.xml")
   if(file.exists("Maize.out")) file.remove("Maize.out")
   if(file.exists("Maize.sum")) file.remove("Maize.sum")
+}
+
+## Testing the optimization of a single parameter 
+## comparing methods
+if(FALSE){
+  
+  tmp <- tempdir()
+  setwd(tmp)
+  
+  extd.dir <- system.file("extdata", package = "apsimx")
+  
+  file.copy(file.path(extd.dir, "Ames.met"), ".")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  
+  sim0 <- apsimx("Wheat-opt-ex.apsimx")
+
+  pp1 <- "Wheat.Leaf.Photosynthesis.RUE.FixedValue"
+  pp2 <- "Wheat.Cultivars.USA.Yecora.BasePhyllochron"
+  
+  ## Testing the "unreliable" method
+  start <- Sys.time()
+  wop.1 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                        parm.paths = pp1,
+                        data = obsWheat, 
+                        replacement = TRUE,
+                        initial.values = 1.2)
+  end <- Sys.time() ## It took 3.25 minutes
+  
+  ## Plus hessian
+  file.remove("Wheat-opt-ex.apsimx")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  
+  start <- Sys.time()
+  wop.h.1 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                          parm.paths = pp1,
+                          data = obsWheat, 
+                          replacement = TRUE,
+                          hessian = TRUE,
+                          initial.values = 1.2)
+  end <- Sys.time() ## It took 3.4 minutes
+ 
+  ## Erase and try Brent 
+  file.remove("Wheat-opt-ex.apsimx")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+ 
+  start <- Sys.time()
+  wop.b.1 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                          parm.paths = pp1,
+                          data = obsWheat, 
+                          method = "Brent",
+                          lower = 0.5, upper = 2,
+                          replacement = TRUE,
+                          initial.values = 1.2)
+  end <- Sys.time() ## It took 2.18 minutes 
+  ## Brent, naturally, will not provide the best solution if
+  ## the solution is outside the interval (lower, upper)
+  
+  ## Erase and try L-BFGS-B
+  file.remove("Wheat-opt-ex.apsimx")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  
+  start <- Sys.time()
+  wop.bfgs.1 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                          parm.paths = pp1,
+                          data = obsWheat, 
+                          method = "L-BFGS-B",
+                          lower = 0.5,
+                          hessian = TRUE,
+                          replacement = TRUE,
+                          initial.values = 1.2,
+                          control = list(trace = 3))
+  end <- Sys.time() ## It took 7.02 minutes 
+  
+  ## Next step is to test whether weighting affects the estimate of the vcov
+  ## and how different it is to the Bayesian analysis
+  
+  ## Erase and try L-BFGS-B
+  file.remove("Wheat-opt-ex.apsimx")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  
+  start <- Sys.time()
+  wop.bfgs.2 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                             parm.paths = pp1,
+                             data = obsWheat,
+                             weights = "mean",
+                             method = "L-BFGS-B",
+                             lower = 0.5,
+                             hessian = TRUE,
+                             replacement = TRUE,
+                             initial.values = 1.2,
+                             control = list(trace = 3))
+  end <- Sys.time() ## It took 4.16 minutes 
+  
+  ## Testing with two parameters
+  file.remove("Wheat-opt-ex.apsimx")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  ## Testing with two parameters
+  start <- Sys.time()
+  wop.2 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                        parm.paths = c(pp1, pp2),
+                        data = obsWheat, 
+                        replacement = c(TRUE, TRUE),
+                        hessian = TRUE,
+                        initial.values = c(1.2, 120))
+  end <- Sys.time() ## It took 15.7 minutes
+  
+  ## Testing with two parameters, without the hessian
+  file.remove("Wheat-opt-ex.apsimx")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  ## Testing with two parameters
+  start <- Sys.time()
+  wop.nh.2 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                        parm.paths = c(pp1, pp2),
+                        data = obsWheat, 
+                        replacement = c(TRUE, TRUE),
+                        initial.values = c(1.2, 120))
+  end <- Sys.time() ## It took 13.97 minutes
+  
+  file.remove("Wheat-opt-ex.apsimx")
+  file.copy(file.path(extd.dir, "Wheat-opt-ex.apsimx"), ".")
+  ## Testing with two parameters
+  start <- Sys.time()
+  wop.3 <- optim_apsimx("Wheat-opt-ex.apsimx", 
+                        parm.paths = c(pp1, pp2),
+                        data = obsWheat, 
+                        type = "ucminf",
+                        weights = "mean",
+                        replacement = c(TRUE, TRUE),
+                        hessian = TRUE,
+                        initial.values = c(1.2, 120))
+  end <- Sys.time() ## It took 28.97 minutes and it did not find the right answer
 }